make code more compact
