+++ webcam-capture-drivers/driver-raspberrypi/src/main/java/com/github/sarxos/webcam/ds/raspberrypi/RaspistillDevice.java
@@ -18,7 +18,7 @@ public class RaspistillDevice extends IPCDevice {
 	 */
 	private final static char CAPTRE_TRIGGER_INPUT = '\n';
 	private final static char CAPTRE_TERMINTE_INPUT = 'x';
-
+	private PNGDecoder decoder=null;
 	/**
 	 * Creates a new instance of RaspistillDevice.
 	 * 
@@ -40,7 +40,7 @@ public class RaspistillDevice extends IPCDevice {
 			e.printStackTrace();
 		}
 	}
-
+	
 	@Override
 	protected Runnable newErrorConsumeWorker() {
 		return new ErrorConsumeWorker();
@@ -64,11 +64,18 @@ public class RaspistillDevice extends IPCDevice {
 
 	@Override
 	public BufferedImage getImage() {
+		if(decoder==null) {
+			try {
+				decoder=new PNGDecoder(in);
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+		}
+		
 		try {
 			out.write(CAPTRE_TRIGGER_INPUT);
 			out.flush();
-			BufferedImage frame = new PNGDecoder(in).decode();
-			in.skip(16);// each time of decode, there will be 16 bytes should be skipped
+			BufferedImage frame = decoder.nextFrame();
 			return frame;
 		} catch (IOException e) {
 			e.printStackTrace();
